Fix delete button
